+++ addOns/authhelper/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - Add wait authentication step to Browser Based Authentication.
+- Include Web Element's selector in the Authentication Report.
 
 ### Fixed
 - Do not fail the authentication on diagnostic errors.